improve code sharing
