+++ hyrisecockpit/frontend/tests/e2e/specs/metrics/access.spec.ts
@@ -1,84 +1,27 @@
-import { useBackendMock } from "../../setup/backendMock";
 import { generateRandomInt } from "../../setup/helpers";
-import { getRoute } from "../views/helpers";
+import { assertHeatMapData, getSelector, getSelectorWithID } from "./helpers";
+import { clickContentOfSelect } from "../helpers";
 import {
-  getSelector,
-  getSelectorWithID,
-  getDetailsSelectorWithID,
-  assertHeatMapData,
-} from "./helpers";
-import { clickContentOfSelect, waitForChartRender } from "../helpers";
+  testNoVisibilityOnOverview,
+  testGenericChartsOnPage,
+} from "./abstractTests";
 
-const backend = useBackendMock();
+const metric = "secondAccess";
+const request = "chunks";
+const layout = { xtitle: "Columns", ytitle: "Chunks" };
+const transform = (xhr: any): any => {
+  return xhr.response.body.body.chunks_data;
+};
 
-let databases: any[] = [];
-let data: any = {};
-let items: any = [];
-let index: any = 0;
+/* test no existence on overview */
+testNoVisibilityOnOverview(metric);
 
-// test on overview
-describe("visiting the overview page", () => {
-  before(() => {
-    cy.setupAppState(backend).then((xhr: any) => {
-      databases = xhr.response.body;
-    });
-    cy.visit(getRoute("overview"));
-    waitForChartRender();
-  });
-
-  // test chart
-  it("will not show a chart", () => {
-    cy.get(getSelector("secondAccess")).should("not.exist");
-  });
-
-  // test details
-  it("will not show metric details", () => {
-    databases.forEach((database: any) => {
-      cy.get(getDetailsSelectorWithID("secondAccess", database.id)).should(
-        "not.exist"
-      );
-    });
-  });
-});
-
-// test on comparison
-describe("visiting the comparison page", () => {
-  before(() => {
-    cy.setupAppState(backend).then((xhr: any) => {
-      databases = xhr.response.body;
-    });
-    cy.visit(getRoute("comparison"));
-    cy.setupData("chunks").then((xhr: any) => {
-      data = xhr.response.body.body.chunks_data;
-    });
-    waitForChartRender();
-  });
-
-  // test layout
-  it("will show the correct titles", () => {
-    databases.forEach((database: any) => {
-      cy.get(getSelectorWithID("secondAccess", database.id)).should(
-        (elements: any) => {
-          const layout = elements[0].layout;
-          expect(layout.xaxis.title.text).to.eq("Columns");
-          expect(layout.yaxis.title.text).to.eq("Chunks");
-        }
-      );
-    });
-  });
-
-  // test details
-  it("will not show metric details", () => {
-    databases.forEach((database: any) => {
-      cy.get(getDetailsSelectorWithID("secondAccess", database.id)).should(
-        "not.exist"
-      );
-    });
-  });
-
+/* test generic and custom on comparison */
+testGenericChartsOnPage({ metric, request }, layout, transform, (getData) => {
   // test data with no selected table
   describe("when no table is selected", () => {
     it("will show the correct empty metric data", () => {
+      const { databases } = getData();
       databases.forEach((database: any) => {
         cy.get(getSelectorWithID("secondAccess", database.id)).should(
           (elements: any) => {
@@ -92,8 +35,9 @@ describe("visiting the comparison page", () => {
   // test data with selected table
   describe("when a table is selected", () => {
     it("will show the correct metric data", () => {
-      items = Object.keys(data[databases[0].id]);
-      index = generateRandomInt(0, items.length);
+      const { data, databases } = getData();
+      const items = Object.keys(data[databases[0].id]);
+      const index = generateRandomInt(0, items.length);
 
       databases.forEach((database: any, idx: number) => {
         cy.get(getSelector("secondAccessSelect"))
@@ -117,6 +61,10 @@ describe("visiting the comparison page", () => {
   // test detailed view
   describe("when clicking the detailded view button", () => {
     it("will open a detailed heatmap view", () => {
+      const { data, databases } = getData();
+      const items = Object.keys(data[databases[0].id]);
+      const index = generateRandomInt(0, items.length);
+
       databases.forEach((database: any) => {
         cy.get(getSelectorWithID("firstAccess", database.id)).should(
           "not.exist"
