+++ hyrisecockpit/frontend/tests/e2e/specs/metrics/cpu.spec.ts
@@ -1,10 +1,13 @@
-import { testLinechartOnComparison, testLinechartOnOverview } from "./helpers";
+import {
+  testLinechartOnComparison,
+  testLinechartOnOverview,
+} from "./abstractTests";
 
 const metric = "cpu";
 const request = "system";
 const layout = { title: "Workload in %", min: 0, max: 105 };
-const transform = (data: any, xhr: any): any => {
-  data = {};
+const transform = (xhr: any): any => {
+  const data: any = {};
   xhr.response.body.forEach((entry: any) => {
     data[entry.id] = entry.system_data[0].system_data.cpu.cpu_process_usage;
   });
