+++ hyrisecockpit/frontend/tests/e2e/specs/metrics/memoryFootprint.spec.ts
@@ -1,11 +1,14 @@
-import { testLinechartOnComparison, testLinechartOnOverview } from "./helpers";
+import {
+  testLinechartOnComparison,
+  testLinechartOnOverview,
+} from "./abstractTests";
 import { getDatabaseMemoryFootprint } from "../databases/helpers";
 
 const metric = "memoryFootprint";
 const request = "storage";
 const layout = { title: "Memory Footprint in MB", min: 0 };
-const transform = (data: any, xhr: any): any => {
-  data = {};
+const transform = (xhr: any): any => {
+  const data: any = {};
   Object.entries(xhr.response.body.body.storage).forEach(([key, entry]) => {
     data[key] = getDatabaseMemoryFootprint(entry);
   });
