+++ hyrisecockpit/frontend/tests/e2e/specs/metrics/queueLength.spec.ts
@@ -1,10 +1,13 @@
-import { testLinechartOnComparison, testLinechartOnOverview } from "./helpers";
+import {
+  testLinechartOnComparison,
+  testLinechartOnOverview,
+} from "./abstractTests";
 import { assignToObject } from "../helpers";
 
 const metric = "queueLength";
 const request = "queue_length";
 const layout = { title: "Number of queries", min: 0 };
-const transform = (data: any, xhr: any): any => {
+const transform = (xhr: any): any => {
   return assignToObject(xhr.response.body, "queue_length");
 };
