+++ hyrisecockpit/frontend/tests/e2e/specs/metrics/throughput.spec.ts
@@ -1,10 +1,13 @@
-import { testLinechartOnComparison, testLinechartOnOverview } from "./helpers";
+import {
+  testLinechartOnComparison,
+  testLinechartOnOverview,
+} from "./abstractTests";
 import { assignToObject } from "../helpers";
 
 const metric = "throughput";
 const request = "throughput";
 const layout = { title: "Number of queries / s", min: 0 };
-const transform = (data: any, xhr: any): any => {
+const transform = (xhr: any): any => {
   return assignToObject(xhr.response.body, "throughput");
 };
