+++ hyrisecockpit/frontend/tests/e2e/specs/views/navigation.spec.ts
@@ -1,5 +1,9 @@
-import { testElementTrigger } from "../abstractTests";
-import { routes, selectors, testRedirection } from "./helpers";
+import {
+  routes,
+  selectors,
+  testElementTrigger,
+  testRedirection,
+} from "./helpers";
 
 describe("Change page routes", () => {
   beforeEach(() => {
